requirements.txt
streamlit
yfinance
tensorflow
scikit-learn
vaderSentiment
newsapi-python
numpy
pandas